move FilterByNames to its own folder
